Pull out helpers from authenticate and prestop
